[14.0][MIG] account_invoice_margin_sale: Migration to 14.0
