--- account_invoice_margin/README.rst
@@ -1,110 +1,0 @@
-======================
-Account Invoice Margin
-======================
-
-.. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! This file is generated by oca-gen-addon-readme !!
-   !! changes will be overwritten.                   !!
-   !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-
-.. |badge1| image:: https://img.shields.io/badge/maturity-Production%2FStable-green.png
-    :target: https://odoo-community.org/page/development-status
-    :alt: Production/Stable
-.. |badge2| image:: https://img.shields.io/badge/licence-AGPL--3-blue.png
-    :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
-    :alt: License: AGPL-3
-.. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fmargin--analysis-lightgray.png?logo=github
-    :target: https://github.com/OCA/margin-analysis/tree/13.0/account_invoice_margin
-    :alt: OCA/margin-analysis
-.. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/margin-analysis-13-0/margin-analysis-13-0-account_invoice_margin
-    :alt: Translate me on Weblate
-.. |badge5| image:: https://img.shields.io/badge/runbot-Try%20me-875A7B.png
-    :target: https://runbot.odoo-community.org/runbot/132/13.0
-    :alt: Try me on Runbot
-
-|badge1| |badge2| |badge3| |badge4| |badge5| 
-
-This module allows to show sales margin in invoices.
-
-The new information is available:
-
-* on the account form view
-
-.. figure:: https://raw.githubusercontent.com/OCA/margin-analysis/13.0/account_invoice_margin/static/description/account_invoice_form.png
-
-* and on the account tree view
-
-.. figure:: https://raw.githubusercontent.com/OCA/margin-analysis/13.0/account_invoice_margin/static/description/account_invoice_tree.png
-
-**Table of contents**
-
-.. contents::
-   :local:
-
-Usage
-=====
-
-This module also adds a security group.
-
-#. To activate it go to user and active "Show Invoice Margin" in
-   security options.
-
-Bug Tracker
-===========
-
-Bugs are tracked on `GitHub Issues <https://github.com/OCA/margin-analysis/issues>`_.
-In case of trouble, please check there if your issue has already been reported.
-If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/margin-analysis/issues/new?body=module:%20account_invoice_margin%0Aversion:%2013.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
-
-Do not contact contributors directly about support or help with technical issues.
-
-Credits
-=======
-
-Authors
-~~~~~~~
-
-* Tecnativa
-* GRAP
-
-Contributors
-~~~~~~~~~~~~
-
-* `Tecnativa <https://www.tecnativa.com>`__:
-
-  * Sergio Teruel
-
-* `GRAP <http://www.grap.coop>`__:
-
-  * Sylvain LE GAL (https://twitter.com/legalsylvain)
-
-* `Open Source Integrators <https://www.opensourceintegrators.com>`__:
-
-  * Bhavesh Odedra
-
-Maintainers
-~~~~~~~~~~~
-
-This module is maintained by the OCA.
-
-.. image:: https://odoo-community.org/logo.png
-   :alt: Odoo Community Association
-   :target: https://odoo-community.org
-
-OCA, or the Odoo Community Association, is a nonprofit organization whose
-mission is to support the collaborative development of Odoo features and
-promote its widespread use.
-
-.. |maintainer-sergio-teruel| image:: https://github.com/sergio-teruel.png?size=40px
-    :target: https://github.com/sergio-teruel
-    :alt: sergio-teruel
-
-Current `maintainer <https://odoo-community.org/page/maintainer-role>`__:
-
-|maintainer-sergio-teruel| 
-
-This module is part of the `OCA/margin-analysis <https://github.com/OCA/margin-analysis/tree/13.0/account_invoice_margin>`_ project on GitHub.
-
-You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
