--- account_invoice_margin/__init__.py
@@ -1,5 +1,0 @@
-# © 2017 Sergio Teruel <sergio.teruel@tecnativa.com>
-# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
-
-from . import models
-from . import report
