--- account_invoice_margin/i18n/es.po
@@ -1,78 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* account_invoice_margin
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 9.0c\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-12-04 13:11+0000\n"
-"PO-Revision-Date: 2020-12-16 22:19+0000\n"
-"Last-Translator: Harald Panten <harald.panten@sygel.es>\n"
-"Language-Team: \n"
-"Language: es\n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: nplurals=2; plural=n != 1;\n"
-"X-Generator: Weblate 4.3.2\n"
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__purchase_price
-msgid "Cost"
-msgstr "Coste"
-
-#. module: account_invoice_margin
-#: model:ir.model,name:account_invoice_margin.model_account_invoice_report
-msgid "Invoices Statistics"
-msgstr "Estadísticas de facturas"
-
-#. module: account_invoice_margin
-#: model:ir.model,name:account_invoice_margin.model_account_move
-msgid "Journal Entries"
-msgstr "Asientos Contables"
-
-#. module: account_invoice_margin
-#: model:ir.model,name:account_invoice_margin.model_account_move_line
-msgid "Journal Item"
-msgstr "Apunte Contable"
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_invoice_report__margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move__margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__margin
-msgid "Margin"
-msgstr "Margen"
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move__margin_percent
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__margin_percent
-msgid "Margin (%)"
-msgstr "Margen (%)"
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move__margin_signed
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__margin_signed
-msgid "Margin Signed"
-msgstr "Margen firmado"
-
-#. module: account_invoice_margin
-#: model:res.groups,name:account_invoice_margin.group_account_invoice_margin_security
-msgid "Show Invoice Margin"
-msgstr "Mostrar margen en facturas"
-
-#. module: account_invoice_margin
-#: model_terms:ir.ui.view,arch_db:account_invoice_margin.view_account_invoice_tree
-msgid "Total"
-msgstr "Total"
-
-#, fuzzy
-#~| msgid "Invoice Line"
-#~ msgid "Invoice"
-#~ msgstr "Línea factura"
-
-#~ msgid "Invoice Line"
-#~ msgstr "Línea factura"
-
-#~ msgid "% Margin"
-#~ msgstr "% Margen"
