--- account_invoice_margin/i18n/fr.po
@@ -1,72 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-#   * account_invoice_margin
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 12.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2019-12-11 17:47+0000\n"
-"PO-Revision-Date: 2019-12-11 17:47+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"Language: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__purchase_price
-msgid "Cost"
-msgstr "Coût"
-
-#. module: account_invoice_margin
-#: model:ir.model,name:account_invoice_margin.model_account_invoice_report
-msgid "Invoices Statistics"
-msgstr "Statistiques des factures"
-
-#. module: account_invoice_margin
-#: model:ir.model,name:account_invoice_margin.model_account_move
-msgid "Journal Entries"
-msgstr ""
-
-#. module: account_invoice_margin
-#: model:ir.model,name:account_invoice_margin.model_account_move_line
-msgid "Journal Item"
-msgstr ""
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_invoice_report__margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move__margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__margin
-msgid "Margin"
-msgstr "Marge"
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move__margin_percent
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__margin_percent
-msgid "Margin (%)"
-msgstr "Marge (%)"
-
-#. module: account_invoice_margin
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move__margin_signed
-#: model:ir.model.fields,field_description:account_invoice_margin.field_account_move_line__margin_signed
-msgid "Margin Signed"
-msgstr "Marge"
-
-#. module: account_invoice_margin
-#: model:res.groups,name:account_invoice_margin.group_account_invoice_margin_security
-msgid "Show Invoice Margin"
-msgstr "Voir la marge sur les factures"
-
-#. module: account_invoice_margin
-#: model_terms:ir.ui.view,arch_db:account_invoice_margin.view_account_invoice_tree
-msgid "Total"
-msgstr "Total"
-
-#~ msgid "Invoice"
-#~ msgstr "Facture"
-
-#~ msgid "Invoice Line"
-#~ msgstr "Ligne de facture"
