--- account_invoice_margin/models/__init__.py
@@ -1,4 +1,0 @@
-# © 2017 Sergio Teruel <sergio.teruel@tecnativa.com>
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
-from . import account_invoice
