--- account_invoice_margin/readme/CONTRIBUTORS.rst
@@ -1,11 +1,0 @@
-* `Tecnativa <https://www.tecnativa.com>`__:
-
-  * Sergio Teruel
-
-* `GRAP <http://www.grap.coop>`__:
-
-  * Sylvain LE GAL (https://twitter.com/legalsylvain)
-
-* `Open Source Integrators <https://www.opensourceintegrators.com>`__:
-
-  * Bhavesh Odedra
