--- account_invoice_margin/readme/DESCRIPTION.rst
@@ -1,11 +1,0 @@
-This module allows to show sales margin in invoices.
-
-The new information is available:
-
-* on the account form view
-
-.. figure:: ../static/description/account_invoice_form.png
-
-* and on the account tree view
-
-.. figure:: ../static/description/account_invoice_tree.png
