--- account_invoice_margin/readme/USAGE.rst
@@ -1,4 +1,0 @@
-This module also adds a security group.
-
-#. To activate it go to user and active "Show Invoice Margin" in
-   security options.
