--- account_invoice_margin/report/__init__.py
@@ -1,1 +1,0 @@
-from . import account_invoice_report
