--- account_invoice_margin/security/account_invoice_margin_security.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- Copyright 2017 Sergio Teruel - Tecnativa <sergio.teruel@tecnativa.com>
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl). -->
-<odoo>
-    <record id="group_account_invoice_margin_security" model="res.groups">
-        <field name="name">Show Invoice Margin</field>
-        <field name="category_id" ref="base.module_category_hidden" />
-        <field name="users" eval="[(4, ref('base.user_admin'))]" />
-    </record>
-</odoo>
