--- account_invoice_margin/views/account_invoice_margin_view.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<!-- © 2017 Sergio Teruel <sergio.teruel@tecnativa.com>
-     License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl-3). -->
-<odoo>
-    <record id="invoice_margin_form_tree" model="ir.ui.view">
-        <field name="model">account.move</field>
-        <field name="inherit_id" ref="account.view_move_form" />
-        <field
-            name="groups_id"
-            eval="[(4, ref('account_invoice_margin.group_account_invoice_margin_security'))]"
-        />
-        <field name="arch" type="xml">
-            <!-- To allow to store new fields you must add this fields to account move lines view -->
-            <xpath
-                expr="//page[@id='aml_tab']//field[@name='price_unit']"
-                position="after"
-            >
-                <field name="purchase_price" invisible="1" />
-                <field name="margin" invisible="1" />
-                <field name="margin_percent" invisible="1" />
-            </xpath>
-            <!-- To display new fields in invoice lines tab  -->
-            <xpath
-                expr="//field[@name='invoice_line_ids']/tree/field[@name='price_unit']"
-                position="after"
-            >
-                <field name="purchase_price" />
-                <field name="margin" />
-                <field name="margin_percent" />
-            </xpath>
-            <!-- Add fields to embed invoice line form view  -->
-            <xpath
-                expr="//page[@id='invoice_tab']//form//field[@name='discount']"
-                position="after"
-            >
-                <field name="purchase_price" />
-                <field name="margin" />
-                <field name="margin_percent" />
-            </xpath>
-            <field name="amount_untaxed" position="before">
-                <field name="margin" />
-                <field name="margin_percent" />
-            </field>
-        </field>
-    </record>
-    <record id="view_account_invoice_tree" model="ir.ui.view">
-        <field name="model">account.move</field>
-        <field name="inherit_id" ref="account.view_invoice_tree" />
-        <field
-            name="groups_id"
-            eval="[(4, ref('account_invoice_margin.group_account_invoice_margin_security'))]"
-        />
-        <field name="arch" type="xml">
-            <field name="amount_untaxed_signed" position="before">
-                <field name="margin_signed" sum="Total" />
-                <field name="margin_percent" />
-            </field>
-        </field>
-    </record>
-</odoo>
