+++ account_invoice_margin_sale/__init__.py
@@ -1,3 +1,3 @@
-# License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from . import models
