+++ account_invoice_margin_sale/models/sale.py
@@ -8,7 +8,7 @@ from odoo import models
 class SaleOrderLine(models.Model):
     _inherit = "sale.order.line"
 
-    def _prepare_invoice_line(self):
-        vals = super()._prepare_invoice_line()
+    def _prepare_invoice_line(self, **optional_values):
+        vals = super()._prepare_invoice_line(**optional_values)
         vals["purchase_price"] = self.purchase_price
         return vals
