+++ account_invoice_margin_sale/tests/test_account_invoice_margin_sale.py
@@ -47,7 +47,6 @@ class TestAccountInvoiceMargin(SavepointCase):
                 "standard_price": 100.00,
             }
         )
-        cls.product.property_account_receivable_id = cls.account
         pricelist = cls.env["product.pricelist"].create({"name": "Public Pricelist"})
 
         cls.sale_order = cls.env["sale.order"].create(
@@ -75,3 +74,57 @@ class TestAccountInvoiceMargin(SavepointCase):
         self.sale_order.order_line.purchase_price = 500.00
         invoice = self.sale_order._create_invoices()
         self.assertAlmostEqual(invoice.invoice_line_ids.purchase_price, 500.00, 2)
+
+    def test_invoice_down_payment(self):
+        SaleAdvancePaymentInv = self.env["sale.advance.payment.inv"]
+        AccountMove = self.env["account.move"]
+        product = self.env["product.product"].create(
+            {
+                "name": "test product for down payment",
+                "categ_id": self.product_categ.id,
+                "uom_id": self.env.ref("uom.product_uom_unit").id,
+                "list_price": 1000.00,
+                "standard_price": 500.00,
+                "type": "service",
+                "invoice_policy": "order",
+            }
+        )
+        self.order = self.env["sale.order"].create(
+            {
+                "partner_id": self.partner.id,
+                "order_line": [
+                    (
+                        0,
+                        False,
+                        {
+                            "product_id": product.id,
+                            "name": "Testing Product",
+                            "product_uom_qty": 1,
+                            "product_uom": product.uom_id.id,
+                            "price_unit": 1000.00,
+                            "purchase_price": 500.00,
+                        },
+                    ),
+                ],
+            }
+        )
+        self.order.action_confirm()
+        # Create one down payment
+        wiz = SaleAdvancePaymentInv.with_context(
+            active_ids=self.order.ids,
+            open_invoices=True,
+        ).create({"advance_payment_method": "fixed", "fixed_amount": 100.00})
+        action = wiz.create_invoices()
+        invoice_id = action["res_id"]
+        invoice1 = AccountMove.browse(invoice_id)
+        self.assertEqual(invoice1.margin, 0.0)
+
+        # Create regular invoice which has a down payment
+        wiz = SaleAdvancePaymentInv.with_context(
+            active_ids=self.order.ids,
+            open_invoices=True,
+        ).create({"advance_payment_method": "delivered"})
+        wiz.create_invoices()
+        invoice2 = self.order.invoice_ids - invoice1
+        self.assertEqual(invoice2.margin, 500.00)
+        self.assertEqual(invoice2.margin_percent, 50.0)
